+++ source/isaaclab_tasks/isaaclab_tasks/contrib/lift/config/openarm/__init__.py
@@ -16,7 +16,7 @@ from . import agents
 ##
 
 gym.register(
-    id="Isaac-Lift-Cube-OpenArm-v0",
+    id="IsaacContrib-Lift-Cube-OpenArm",
     entry_point="isaaclab.envs:ManagerBasedRLEnv",
     kwargs={
         "env_cfg_entry_point": f"{__name__}.joint_pos_env_cfg:OpenArmCubeLiftEnvCfg",
@@ -27,7 +27,7 @@ gym.register(
 )
 
 gym.register(
-    id="Isaac-Lift-Cube-OpenArm-Play-v0",
+    id="IsaacContrib-Lift-Cube-OpenArm-Play",
     entry_point="isaaclab.envs:ManagerBasedRLEnv",
     kwargs={
         "env_cfg_entry_point": f"{__name__}.joint_pos_env_cfg:OpenArmCubeLiftEnvCfg_PLAY",