+++ source/isaaclab_tasks/isaaclab_tasks/contrib/reach/config/openarm/bimanual/__init__.py
@@ -16,7 +16,7 @@ from . import agents
 ##
 
 gym.register(
-    id="Isaac-Reach-OpenArmBi-v0",
+    id="IsaacContrib-Reach-OpenArmBi",
     entry_point="isaaclab.envs:ManagerBasedRLEnv",
     disable_env_checker=True,
     kwargs={
@@ -27,7 +27,7 @@ gym.register(
 )
 
 gym.register(
-    id="Isaac-Reach-OpenArmBi-Play-v0",
+    id="IsaacContrib-Reach-OpenArmBi-Play",
     entry_point="isaaclab.envs:ManagerBasedRLEnv",
     disable_env_checker=True,
     kwargs={